requirements.txt
# Core dependencies
python-dotenv==1.0.1
PyGithub==2.2.0
chromadb==0.4.22
sentence-transformers==2.5.1
tqdm==4.66.2

# LLM dependencies
openai==1.68.2
anthropic==0.49.0

# Code embeddings and analysis
transformers==4.38.2
torch==2.2.1
accelerate==0.31.0
scikit-learn==1.6.1
numpy==1.26.4
scipy==1.12.0
joblib==1.3.2
regex==2023.12.25
tokenizers==0.15.2

# Vector storage
hnswlib==0.7.0
faiss-cpu==1.7.4

# Utility dependencies
pathlib==1.0.1
typing-extensions==4.11.0
ujson==5.9.0
dataclasses-json==0.6.4
pydantic==2.7.4
psutil==5.9.8

# Development dependencies
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 